+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [...]
 - **[NEW]** Add `shadowed` prop `MeetingPointIcon`
 - **[UPDATE]** Changed `MeetingPointIcon` active display
+- **[UPDATE]** Add `highlightRoad` prop to `Itinerary`
 
 # v18.0.2 (24/01/2020)